.gstack/experiments/phase-minus-1-drfx/pyproject.toml
[project]
name = "phase-minus-1-drfx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.96.0",
    "ccxt>=4.5.49",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "pynescript>=0.3.0",
    "pynesys-pynecore",
    "python-dotenv>=1.2.2",
    "typer>=0.24.1",
]

[tool.uv.sources]
pynesys-pynecore = { git = "https://github.com/PyneSys/pynecore.git" }
